test1.1

+++ html/static/js/user/addengineer.js
@@ -90,7 +90,7 @@ $("#submit").click(function(event) {
   a.number = id;
   a.address=address;
   a.birthDate=buquan(year,2)+'-'+buquan(month,2)+'-'+buquan(day,2);
-  a.baseSalary=parseInt(salary);
+  a.baseSalary=parseFloat(salary);
   a.hometown=domicile;
   a.edu=parseInt(education);
   //console.log(sex);
